볼링 게임 점수 계산
- frame 을 통해 그리는 볼링판 구현
