+++ src/main/java/com/jklee/cleancode/bowling/refactoring/Score.java
@@ -12,12 +12,19 @@ public class Score {
 	private int score;
 	private int leftCount;
 	private int extraLeftCount;
+	private String pointSign;
 
 	public static Score init(int point) {
 		return Score.builder()
 				.score(point)
 				.extraLeftCount(point == 10 ? 2 : 1)
 				.leftCount(point == 10 ? 2 : 1)
+				.pointSign(point == 10
+						? "X"
+						: point == 0
+								? "-|"
+								: point + "|"
+				)
 				.build();
 	}
 
@@ -26,6 +33,12 @@ public class Score {
 			score += point;
 			leftCount++;
 			extraLeftCount = score == 10 ? 1 : 0;
+
+			pointSign += leftCount == 2 && score == 10
+					? "/"
+					: point == 0
+							? "-"
+							: String.valueOf(point);
 
 			return true;
 		}
